Add new `gix-revwalk` crate for support types related to revision walking.
That way, they can be more selectively used without pulling in all of `gix-revision`
