+++ Cargo.toml
@@ -250,6 +250,7 @@ members = [
     "gix-tui",
     "gix-tix",
     "gix-archive",
+    "gix-revwalk",
 
     "cargo-smart-release",
     "tests/tools",
