+++ gix-revision/Cargo.toml
@@ -21,17 +21,17 @@ gix-hash = { version = "^0.11.2", path = "../gix-hash" }
 gix-object = { version = "^0.30.0", path = "../gix-object" }
 gix-date = { version = "^0.5.1", path = "../gix-date" }
 gix-hashtable = { version = "^0.2.1", path = "../gix-hashtable" }
-gix-commitgraph = { version = "^0.16.0", path = "../gix-commitgraph" }
+gix-revwalk = { version = "^0.1.0", path = "../gix-revwalk" }
 
 bstr = { version = "1.3.0", default-features = false, features = ["std"]}
 thiserror = "1.0.26"
 serde = { version = "1.0.114", optional = true, default-features = false, features = ["derive"] }
-smallvec = "1.10.0"
 document-features = { version = "0.2.1", optional = true }
 
 [dev-dependencies]
 gix-odb = { path = "../gix-odb" }
 gix-testtools = { path = "../tests/tools" }
+gix-commitgraph = { path = "../gix-commitgraph" }
 
 [package.metadata.docs.rs]
 all-features = true
