+++ gix-revision/tests/revision.rs
@@ -1,5 +1,4 @@
 mod describe;
-mod graph;
 mod spec;
 pub type Result<T = ()> = std::result::Result<T, Box<dyn std::error::Error + 'static>>;
