+++ gix-revwalk/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "gix-revwalk"
+version = "0.1.0"
+repository = "https://github.com/Byron/gitoxide"
+license = "MIT/Apache-2.0"
+description = "A crate providing utilities for walking the revision graph"
+authors = ["Sebastian Thiel <sebastian.thiel@icloud.com>"]
+edition = "2021"
+include = ["src/**/*", "LICENSE-*", "CHANGELOG.md"]
+rust-version = "1.64"
+
+[lib]
+doctest = false
+
+[dependencies]
+gix-hash = { version = "^0.11.2", path = "../gix-hash" }
+gix-object = { version = "^0.30.0", path = "../gix-object" }
+gix-hashtable = { version = "^0.2.1", path = "../gix-hashtable" }
+gix-commitgraph = { version = "^0.16.0", path = "../gix-commitgraph" }
+
+thiserror = "1.0.26"
+smallvec = "1.10.0"
