+++ gix-revwalk/src/graph/errors.rs
